refactor(rust): moved the builder functions to their respective structs
