+++ implementations/rust/ockam/ockam_api/tests/credential_issuer.rs
@@ -2,14 +2,12 @@ use ockam_core::compat::collections::{BTreeMap, HashMap};
 use ockam_core::compat::sync::Arc;
 
 use ockam::identity::credential::Timestamp;
-use ockam::identity::{identities, secure_channels, AttributesEntry, IdentitiesStorage};
+use ockam::identity::{identities, AttributesEntry, IdentitiesStorage};
 use ockam::route;
 use ockam_api::bootstrapped_identities_store::{BootstrapedIdentityStore, PreTrustedIdentities};
 use ockam_core::compat::rand::random_string;
 use ockam_core::{AllowAll, Result};
-use ockam_identity::{
-    CredentialsIssuer, CredentialsIssuerClient, SecureChannelListenerOptions, SecureChannelOptions,
-};
+use ockam_identity::{CredentialsIssuer, CredentialsIssuerClient, Identities, SecureChannelListenerOptions, SecureChannelOptions, SecureChannels};
 use ockam_node::Context;
 
 #[ockam_macros::test]
@@ -42,11 +40,11 @@ async fn credential(ctx: &mut Context) -> Result<()> {
     // Now recreate the identities services with the previous vault
     // (so that the authority can verify its signature)
     // and the repository containing the trusted identities
-    let identities = identities::builder()
+    let identities = Identities::builder()
         .with_identities_repository(Arc::new(boostrapped))
         .with_identities_vault(identities.clone().vault())
         .build();
-    let secure_channels = secure_channels::builder()
+    let secure_channels = SecureChannels::builder()
         .with_identities(identities.clone())
         .build();
     let identities_creation = identities.identities_creation();
@@ -65,14 +63,14 @@ async fn credential(ctx: &mut Context) -> Result<()> {
         auth_identity.clone(),
         "project42".into(),
     )
-    .await?;
+        .await?;
     ctx.start_worker(
         &auth_worker_addr,
         auth,
         AllowAll, // In reality there is ABAC rule here.
         AllowAll,
     )
-    .await?;
+        .await?;
 
     // Connect to the API channel from the member:
     let e2a = secure_channels
